Moved CPU tests and re-enabled SIMD test as I'm pretty sure I found the issue (missing clobbar value)
