+++ kernel/UnitTests/AArch64/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(kernel8.elf
+    PRIVATE
+        CPUTests.h CPUTests.cpp
+)
