+++ kernel/UnitTests/AArch64/CPUTests.h
@@ -1,0 +1,12 @@
+#ifndef KERNEL_UNITTESTS_AARCH64_CPUTESTS_H
+#define KERNEL_UNITTESTS_AARCH64_CPUTESTS_H
+
+namespace UnitTests::AArch64::CPU
+{
+    /**
+     * Run all runtime tests
+     */
+    void Run();
+}
+
+#endif // KERNEL_UNITTESTS_AARCH64_CPUTESTS_H
